+++ relayer-cli/src/commands/query.rs
@@ -5,6 +5,7 @@ use abscissa_core::{Command, Options, Runnable};
 mod channel;
 mod client;
 mod connection;
+mod connections;
 mod packet;
 
 /// `query` subcommand
@@ -17,6 +18,10 @@ pub enum QueryCmd {
     /// The `query connection` subcommand
     #[options(help = "query information about connection(s)")]
     Connection(QueryConnectionCmds),
+
+    /// The `query connections` subcommand
+    #[options(help = "query the identifiers of all connection on a chain")]
+    Connections(connections::QueryConnectionsCmd),
 
     /// The `query channel` subcommand
     #[options(help = "query information about channel(s)")]