+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - JSON output for queries and txs ([#500])
   - Implement commands for channel close init and confirm ([#538])
   - Implement command to perform the handshake for a new channel ([#557])
+  - Query all connections CLI ([#553])
 
 - [relayer]
   - Added retry mechanism, restructured relayer ([#519])
@@ -64,6 +65,7 @@
 [#538]: https://github.com/informalsystems/ibc-rs/issues/538
 [#540]: https://github.com/informalsystems/ibc-rs/issues/540
 [#554]: https://github.com/informalsystems/ibc-rs/issues/554
+[#553]: https://github.com/informalsystems/ibc-rs/issues/553
 [#557]: https://github.com/informalsystems/ibc-rs/issues/557
 
 ## v0.0.6